Replaced the param converter by a value resolver
