+++ Tests/Fixtures/ActionArgumentsBundle/ActionArgumentsBundle.php
@@ -1,0 +1,18 @@
+<?php
+
+/*
+ * This file is part of the Symfony framework.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Tests\Fixtures\ActionArgumentsBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class ActionArgumentsBundle extends Bundle
+{
+}
